requirements.txt
# ── Framework ─────────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.34.0

# ── Database ──────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
greenlet==3.1.1

# ── Auth ──────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-multipart==0.0.19

# ── Settings ──────────────────────────────────────────────
pydantic-settings==2.7.1

# ── Testing ───────────────────────────────────────────────
pytest==8.3.4
pytest-asyncio==0.25.3
httpx==0.28.1
aiosqlite==0.20.0

# ── Utilities ─────────────────────────────────────────────
python-dateutil==2.9.0
